Add version system

+++ core/kernel/kernel.c
@@ -598,9 +598,8 @@ int sys_setctx(FsContext ctx)
     return EOK;
 }
 
-/*
- * Syscall: system services
- */
+/*  Syscall: system services */
+
 int sys_info(SysInfo *out)
 {
     uint8_t s0[DISK_SECTOR_SIZE];
@@ -619,7 +618,7 @@ int sys_info(SysInfo *out)
     out->tpa = ((uint32_t)__kernel_base - (uintptr_t)__tpa_base) / 1024;
 
     for (int8_t v = 0; v < MAX_VOLUMES; v++)
-        out->vol_mounted[v] = (volume_run_count((int8_t)v) > 0) ? 1 : 0;
+        out->vol_mounted[v] = volume_run_count(v) > 0 ? 1 : 0;
 
     out->disk_size_kb = disk_block_count();
     out->disk_unalloc_kb = disk_free_blocks();
